Started Repo
Notes.txt just has overarching ideas

vault_condition.txt is the file maintaining the condition values
necessary to declare the state of the vault (level 1..level 5)
It'll be filled in with kosher information once the decision tree
is finished.

condition_to_warning.py simply takes the current condition of the
vault as the input, compares it to the levels found in the
vault_condition.txt and outputs the highest match. Has a --help
flag.
Flags: --infile ~ changes the file being used for input.
       --anymatch ~ this is a switch (on/off, no arguments follow).
            We can either match to a condition level based on hitting
            the minimum of any area, or by hitting the minimum of
            every area. I'm not sure what the final decision will be,
            so this switch will handle that when necessary.