Added Decision_tree.py and image.sh
Image.sh is just a command sequence to draw the graph using graphviz

decision_tree is just a simple directed graph structure with inorder
traversal set up. Next step is to make it read q&a from a txt file,
and place them in the correct nodes.
After that, it needs a malleable evaluation function so it will
direct you to the correct child as you move through the tree, and
then it just needs the real q&a set to work with.

answer can just be passed to <stdin> and image saved, both consumed
by whatever GUI system yalinne writes.