Added tree_from_file.py
Basically runs a parser for a psuedo-language composed only of if and else conditionals
Intended as an easy-to-write decision tree maker for the engies
Reads the tree from input.txt, reads the current values from vals.txt

In no way optimized, runs at O(mn), where n is the number of nodes and m is the number of decisions that have to be made. It's not actually stored as a tree, just a flat array.

works tho